0.4.0 finally, mobs now config from file.

+++ core/src/main/java/com/egoal/darkestpixeldungeon/actors/mobs/npcs/Alchemist.java
@@ -132,7 +132,7 @@ public class Alchemist extends NPC {
   }
 
   @Override
-  public int defenseSkill(Char enemy) {
+  public float defenseSkill(Char enemy) {
     return 1000;
   }
