+++ core/src/main/java/com/egoal/darkestpixeldungeon/actors/mobs/npcs/Ghost.java
@@ -80,7 +80,7 @@ public class Ghost extends NPC {
   }
 
   @Override
-  public int defenseSkill(Char enemy) {
+  public float defenseSkill(Char enemy) {
     return 1000;
   }
