+++ core/src/main/java/com/egoal/darkestpixeldungeon/actors/mobs/npcs/Imp.java
@@ -68,7 +68,7 @@ public class Imp extends NPC {
   }
 
   @Override
-  public int defenseSkill(Char enemy) {
+  public float defenseSkill(Char enemy) {
     return 1000;
   }
