+++ core/src/main/java/com/egoal/darkestpixeldungeon/actors/mobs/npcs/KingStatuary.kt
@@ -98,7 +98,7 @@ class KingStatuary : NPC() {
         return super.act()
     }
 
-    override fun defenseSkill(enemy: Char): Int = 1000
+    override fun defenseSkill(enemy: Char): Float = 1000f
 
     override fun takeDamage(dmg: Damage): Int = 0
