+++ core/src/main/java/com/egoal/darkestpixeldungeon/actors/mobs/npcs/RobotREN.kt
@@ -69,7 +69,7 @@ class RobotREN : NPC() {
         return super.act()
     }
 
-    override fun defenseSkill(enemy: Char): Int = 1000
+    override fun defenseSkill(enemy: Char): Float = 1000f
 
     override fun takeDamage(dmg: Damage): Int = 0
