+++ core/src/main/java/com/egoal/darkestpixeldungeon/actors/mobs/npcs/Statuary.java
@@ -448,7 +448,7 @@ public class Statuary extends NPC {
   }
 
   @Override
-  public int defenseSkill(Char enemy) {
+  public float defenseSkill(Char enemy) {
     return 1000;
   }
