+++ core/src/main/java/com/egoal/darkestpixeldungeon/actors/mobs/npcs/Wandmaker.java
@@ -64,7 +64,7 @@ public class Wandmaker extends NPC {
   }
 
   @Override
-  public int defenseSkill(Char enemy) {
+  public float defenseSkill(Char enemy) {
     return 1000;
   }
